Expose getTimeObject function
